+++ app/Http/Controllers/Dashboard/AwardController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Dashboard\AwardRequest;
-use App\Models\Award;
 use App\Repositories\Classes\AwardRepository;
 use Illuminate\Http\Request;
 
@@ -13,13 +12,7 @@ class AwardController extends Controller
     protected AwardRepository $awardRepository;
     public function __construct(AwardRepository $awardRepository)
     {
-        //        $this->authorize('view_awards',['index']);
-        //        $this->authorize('create_awards', ['create', 'store']);
-        //        $this->authorize('show_awards', ['show']);
-        //        $this->authorize('update_awards', ['update']);
-        //        $this->authorize('delete_awards', ['destroy']);
         $this->awardRepository = $awardRepository;
-
     }
 
     /**
@@ -27,6 +20,7 @@ class AwardController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('view_awards');
         if ($request->ajax()) {
             $awards = $this->awardRepository->findBy($request);
             return response()->json($awards);
@@ -39,6 +33,7 @@ class AwardController extends Controller
      */
     public function create()
     {
+        $this->authorize('create_awards');
         return view('dashboard.awards.create');
     }
 
@@ -47,6 +42,7 @@ class AwardController extends Controller
      */
     public function store(AwardRequest $request)
     {
+        $this->authorize('create_awards');
         $this->awardRepository->store($request->validated());
     }
 
@@ -55,7 +51,7 @@ class AwardController extends Controller
      */
     public function show(string $id)
     {
-        //
+        // $this->authorize('show_awards');
     }
 
     /**
@@ -63,6 +59,7 @@ class AwardController extends Controller
      */
     public function edit(string $id)
     {
+        $this->authorize('update_awards');
         $award = $this->awardRepository->show($id);
         return view('dashboard.awards.edit', compact('award'));
     }
@@ -72,6 +69,7 @@ class AwardController extends Controller
      */
     public function update(AwardRequest $request, string $id)
     {
+        $this->authorize('update_awards');
         $this->awardRepository->update($request->validated(), $id);
     }
 
@@ -80,6 +78,7 @@ class AwardController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete_awards');
         $this->awardRepository->destroy($id);
     }
 }
